Clientservice: Revert default config
My hesitation to introduce default configuration was justified. In this
particular case, the default configuration adds 4 replicas and then we add more
instead of replacing the default replicas. Although this can be fixed in
different ways, I've decided to completely remove default configuration for now
and make all configuration parameters that we expect mandatory. This way,
clientservice won't even start if there is a missing value in the configuration
file. Especially in the current state of the project, this is very helpful so
that you don't have to use GDB and figure out why variables have weird values.