TRS: Use a separate spsc_queue for EventGroupUpdates
This commit adds another spsc_queue that holds updates of type
`kvbc::EventGroupUpdate`. With this all methods associated with
teh spsc_queue of type `kvbc::BlockUpdate` have been replicated
and updated to work with the event_group type. The plan is to
eventually replace the two queues with a single queue that holds
an std::variant.