TRC: Add hash caclucation for event groups
In a prior change, we introduced hash calculation for event groups in the TRS.
With this change we bring the same functionality to TRC. Yes, we are
duplicating code here. This is due to how the code was structured in
closed-source. We will introduce a thin replica hashing library to avoid
duplication later.