Working window refactor (#1725)
* Implemented circular buffer directly into SequenceWithActiveWindow.
Properly handle the situation where we only slide with the size of the Working Window.
Previously we used to clear all Items, which is not necessary.
Move log message for getting from Inactive Window to DEBUG level.

The Working Window consists of the Active Window and optionally
(compile time selected) the Inactive Window. In the general case
when we have both Active and Inactive parts, the date structure
which is a circular buffer looks like this:

                         | firstItemOfActiveWindow
                         v
                           ---------
                   |  |  |/ |  |  | \|  |  |
                 0 |--|--|--|--|--|--|--|--| totalItems
                   |  |  |  |  |  |  |  |  |

To determine the actual Sequence ID of the first element
we use the variable beginningOfActiveWindow. The Inactive
Window if present is always right before the Active Window.
In the example above it occupies elements 6, 7, 0 and 1.
When we scroll the Active Window we have 2 cases:
1. We scroll with more than the size of the Active Window.
   In this case we reset all elements. Even if we could store
   some elements in the Inactive Window, there would be a gap
   between the Active and Inactive Window, therefore we won't
   be able to fully help peers to catch up.
2. We Scroll with less or equal to the size of the Active Window.
   In this case we need to reset as many elements as we scroll
   after the previous (prior to scrolling) last position of the
   Active Window and increment the size of the Inactive Window
   with the same number of elements. It is important to keep in
   mind that the Inactive Window has maximum size, which once
   reached remains constant, unless we clear all elements, thus
   getting it back to zero.

* Remove acquire methods as they are no longer needed by SequenceWithActiveWindow.

Previously we had 2 circular buffers, one for the Active Window and one
for the Inactive Window. This required correct transferring of the content
of the Elements stored in those circular buffers. After the refactor
We store both the Active and the Inactive parts in one circular buffer and
those functions are no longer needed.

* We only clear all elements when we advance with more than the Active Window size.

We need to verify that we properly store the values in the Inactive Window in case
we scroll with size equal to the Active Window size.

* Add tests for Working Window that contains only Active part.
Refactor old tests to cover more cases.

* Added _ to non const member variables.
Const correctness applied to all temporary variables.