BCStateTran: record message arrival time (into handoff)
To be able to track timeouts more precisely, we need to record the exact
time a message has arrived to ST destination. In some cases, I have
noticed that a message can stay up to 600ms in the handoff queue, this
is very inaccurate.
Here we record the message arrival time. In particular, we are
interested in ItemDataMsg messages. We provide a correction while
calling setFetchingTimeStamp when reporting to source selector.