Reconfiguration mechanism improvements (#1738)
* add an option to persist wedge record on wedging commands

* Unwedge command refactoring

1. Add a restart flag to the unwedge command
2. In the unwedge response sign on the epoch number (instead of the wedge point)
3. Remove the unwedge functionality from the ControlStateManager
4. Start a new epoch on unwedge command

* Don't allow metadata removal without starting a new epoch

* Fix a bug with rocksdb client

When copy the rocksdb data, we need to make sure that the WAL directory is updated to the new directory. Otherwise, rocksdb will try to search for WAL in the original location which may cause severe issues

* Add automated restart callback to tester replica

* Fix a bug with restart ready mechainsm

We fix a bug in which a replica gets all restart ready messages prior to the point where it produces its own ready restart message. In this case, the replica won't send a restart ready proof message.
This commit fixes the issue

* Allow a replica to receive restart ready messages even if it is not in wedge point yet

Preventing the reply attack is being done in the onRestartProof (where we restart only if it is valid)

* Send a self restartReadyProof message instead of invoking the handler

This is to reduce the probability of restarting before we completed the async send

* Wedge events refactoring

Refactoring of the wedge event in both, reconfiguration handlers and st_reconfiguration handler
In addition, we handle n-f reset command

* Update Apollo tests with the latest changes

* stabilize the tests by adding a small timeout before shutting down