EventService: Event group start at 1
In the protobuf definition, we are stating that the first event group is 0 but
in fact it is 1. In addition, we've been setting the "first" last verified
state to the requested id and thereby starting at the requested id + 1. The
protobuf documentation for the event service clearly states that the
subscription request will return updates *starting* at the given ID.