TRS: Fix typo in subscription_buffer
This typo induces undefined behavior but made me think more about the wait
function. Given that we can be "stuck" there for a long time, it is a good idea
to check if the stream was cancelled and return early after the wait.
Unfortunately, we cannot interrupt the wait and so the calling thread stays
alive until a new event group comes in. Maybe we should replace wait with
wait_for, check wether we should stop or continue waiting.