Cron: check for pending ticks in the external queue (#1751)
Fix a bug in the Concord Cron's TickGenerator that would allow
generating a tick for a component ID when there is already a pending one
for the same component ID. The path of a tick is the following:
   `internal msg queue -> external msg queue -> consensus`
Make sure that we treat ticks in the external message queue as pending
too instead of only treating ticks that are in consensus as pending.

In order to implement, introduce a set of ticks (per component ID) that
are pending in the external message queue. Attach a callback to ticks in
the external message queue that is called when the tick is processed
(either dropped or consumed for consensus). This callback deletes the
tick from the set of pending ones in the external queue.

Use unordered_set/map instead of ordered ones.

Add support for a callback when a request is popped from the external
queue in the messaging thread in IInternalBFTClient.

Force that TicksGenerator is only created as a shared_ptr as we want to
pass it around to callbacks. It is already used as a shared_ptr anyway.

Add a test case for the new behaviour.