TRS: Don't block on new updates
The gRPC thread which is running this code might be blocked for a long time
depending on the processed requests. If the TRC decides that it is waiting for
too long then it will cancel the subscription and try again. However, the TRS
won't notice this at first because it is blocking and thereby occupying
resources unnecessarily.
This change waits in a loop for a fixed amount of time. Not great, but bettern
than unconditional blocking.