Add check for unique signatures in PreProcessResultMsg validate function
At the moment validatePreProcessResultSignatures() doesn't check if the
signatures attached to the message are from unique replicas. This means
that a malicious primary can send a single signature f+1 times and it
will fool a non-primary replica that the result is valid.

This patch adds check that all signatures are from unique replicas.
SigManager has got public keys for each replica so it is guaranteed that
the message is uniquely signed by f+1 replicas.