Docker update to version 0.28
So far we have installed `clang`, `clang-format-9`, and `clang-tidy-10`. Let's
sync all of them to the same version. We can do further compiler upgrades in a
follow-up. This PR focuses on minimal changes to sync with closed-source.
Note, `clang-tidy-9` has a bug with respect to `if constexpr` statements.
Therefore, we leave it for now.