Adding epoch number to replica's internal message
In this PR, we are adding epoch number to replica's internal messages. Epoch is the lifecycle of a system. Change in epoch number indicates that we started a new system on top of the old
one. For example, on scale-up/down to the cluster, we essentally resatrt the system with new epoch number. So, replicas need to ignore consensus messages from a replica in previous epoch(with some exceptions). Summary of the changes in this PR:
- Added epoch number of the protocol messages
- Added validity check to ignore messages from different epoch number
- Added relevent checks in checkpoint message handler so that replicas in previous epoch can get states via ST