BCStateTran: change the way destination handles lastInBatch flag
lastInBatch may arrive in an ItemDataMsg, as a reply to FetchBlocksMsg
or (in one of the next commits) in FetchREsPagesMsg.
To send another message, the destination shouldn't look if it could have
processed a whole block, but solely on the flag as arrived from the
channel. Furthermore, during GettingMissingResPages, until now, no
message was sent immediately when the flag is up. In case that vBlock
is chunked we may end up sending the next message during onTimerImp().
This is not efficient! Here we fix this inefficiency by calling
processData with lastInBatch flag.