TRS: Transition from legacy events to event groups
Based on the API documentation for event.proto this PR implements logic so that
a legacy event request can receive event groups without interruption.
We throw an exception whenever we try to read a block for legacy events which
actually contains event groups. This exception gets handled in the TRS and
later on, we continue with the subscription stream starting from the first
event group.