Fix non-const lvalue arguments in ThreadPool
Make sure that non-const lvalue arguments are copied as we don't
transport references from async() to the callable. In general, users can
either copy or move arguments. In order to achieve that, we add a
missing std::decay_t when declaring the packaged task parameters type.

Add a comment about intentional lack of support for references. We note,
though, that std::c/ref() can explicitly be used if references are
required.

Add unit tests showing that pointers and non-const lvalues are copied
and that std::ref() works as expected.

Thanks to @arc-vmware for noticing this issue!