BCStateTran: fix wrong assertion
This should assert for the last nextCommittedBlockId_. The reason it has
never failed with the code we had till now, is that usually
nextCommittedBlockId_ == nextRequiredBlock_ when there is timeout, and
in addition, we hardly see bad data from the source. Recently we had
errors coming from RO replica storage layer which caused this assert
to fail. Here we fix it.