TRS: Test transition from legacy to event groups
Introduce tests to guarantee that a legacy event request will transition into
event groups eventually. In order to this, this change adapts the current state
machine implementation. Our API is able to send events or event groups and this
change reflects that. Also, we adjust counters so that we can have events and
event groups in storage at the same time.