TRC: Initialize last verfied event group id
The transition from a legacy event request to event groups revealed a bug in
the TRC. We initialized event groups to its max value which triggered an assert
when receiving event groups. The assert is valid but the initialization can be
changed. In addition, this change renames another variable for readability.