Change type of record_time in EventGroup struct
This commit changes the type of record_time member of EventGroup from
std::chrono::microseconds to google::protobuf::Timestamp in update.hpp.
Without this change, record_time received as type `Timstamp` by the TRC
would be required to changed to type std::chrono::microseconds so as to
be consumed by the event_service, and changed back to type Timestamp
at the event_service, which can potentially cause loss of precision
in the value of record_time.