TRS: Add record_time to event_group
Upon reading, filtering event group updates, the TRS prepares the
event_group object to put on the stream for TRC to consume.
This commit adds record_time to the event_group object.
Note: record_time is generated by the execution engine, it is
defined as the time at which the update was written to the ledger.