Add an unit test with missing Pre-Prepare messages
"adding_pre_prepare_messages_to_status_message" - verifies that the
views manager is able to prevent the replica from moving to higher view
when there is a missing pre-prepare message.
The test also proves that the views manager is capable of filling the status message
of the replica with the necessary information about missing Pre-Prepare messages.