BCStateTran, SourceSelector: Maintain an actual list of sources
Till now, every time a new source is selected, a list of sources (which
isused for cycle end report) is updated immediately.
But if a source does not answer, or if it sends a corrupted data etc.,
it is not really a source.
An actual or "real" source, is one in which at least one block has been
received from (during Gettingmissingblocks or GettingMissingResPages).

Here we move the maintenance of this list into source selector, and
report to it with a new API function whenever a new block is received.