bcstatetransfer_tests: Add a unit test dstValidateRealSourceListReported
This unit test simulates a case in which ST is conducted, 2 sources are
chosen one after another, but none of them answer for multiple
retransmissions. Then a 3rd source is selected and answers, ST is done.
On final callback to inform completion - test validates that the list of
actual sources is just one source - the last one.