KvbAppFilter: Support public event groups
Depending on the tags attached, an event can be either public or private.
At the moment, we store public events but deliver private events only.
This commit solves the problem of consuming both public and private event
groups from storage s.t., event groups are consumed in the order in
which they were generated.

This commit adds a method that reads event group ids from storage
in batches (of default size 10) from the tag table, and generate
an ordered sequence of global event group ids which is eventually used
by the TRS to read event groups from the data table. The index of
this generated sequence determines the event_group_id for the
requesting client.