Remove redundant passing of pre_exec_all in Apollo
This commit removes explicit values for the pre_exec_all parameter in
instantiations of SimpleKVBCProtocol in Apollo where the
SimpleKVBCProtocol is also given an SkvbcTracker. This commit also
modifies SimpleKVBCProtocol.__init__ to accomodate this change by
defaulting the SimpleKVBCProtocol instance's value for pre_exec_all to
the tracker's value of pre_exec_all.

This commit is intended to remove the unnecessary redundancy in
instantiations of SimpleKVBCProtocol that give values for both tracker
and pre_exec_all, as SkvbcTrackers already have a value for
pre_exec_all, and it would be erroneous if a SimpleKVBCProtocol object
was initialized with a tracker and an explicit value for pre_exec_all
not matching the value the tracker already has.