Fetch oldest global event group id from tag-table to getOldestEventGroupBlockId
Event group id 1 will not be the oldest global event group id
after pruning, hence we need to query the tag table to get the latest
global event group id, and use that to fetch the block id associated with
the oldest global event group id.
This method also renames getFirstEventGroupBlockId to
getOldestEventGroupBlockId.