Util: Prometheus wrapper for concord-bft metrics
This change introduces the wrapper from closed source into open-source. The
motivation is that we want to expose a Prometheus port on the clientservice to
query metrics.
Note: Dependencies are always an issue in this repository and hence this change
allows to build other targets without the need for Prometheus. In addition,
instead of adding the Prometheus wrapper to the util library, and thereby
exposing the dependency to all consumers, we create a new util_prometheus
library which is basically util + prometheus wrapper. Why not create a
prometheus_wrapper library without the other utils? I chose not to do that due
to the code organization. The source code is placed inside the util directory
(where it belongs IMHO) and therefore, it is expected to be part of the util
library or a specialized version of it. I'm open to discuss other suggestions.