ConcordClient: Own Basic-/UpdateQueue
So far, the TRC owned the UpdateQueue which creates an implicit dependency to
TRC for the user of concord client because we expose those structures in the
API. With this change, concord client takes ownership of the UpdateQueue a
basic implementation and the type definitions of the updates that will be
pushed on the queue. Thereby, we create a `concordclient-event-api` library
which is used by the TRC. Effectively, the TRC now depends on a portion of
concord client. This is intentional because we want to hide the implementation
of the subscription mechanism in concord client and rather create this
dependency instead.