ConcordClient: User provided update queue
The intention of the UpdateQueue interface is to be implemented by the user of
concord client / thin replica client. This change implements the idea by
requiring the user to provide a queue implementation to the `subscribe` API
call. However, with the prior introduced `concordclient-event-api` library, the
caller has access to a basic update queue implementation.
Implementation detail: The TRC is instantiated with a queue. Meaning, for every
new subscription, we have to create a new TRC object. Looking forward, this is
desired in order to allow for multiple subscriptions because every subscriber
needs their own update queue.