Fixing the data type of maxSpaceForReqs and review of primaryCombinedReqSize
Following 2 issues are handled in this PR
1) The datatype of maxSpaceForReqs is uint_16. This variable
stores the size of max space needed for the PrePrepareMsg.
So this variable can limit the size of PPM to 64KB.
2) The primaryCombinedReqSize variable should keep track of
actual total size of ClientRequestMsg. So every time we push CRM
into PPM, we have to increment it. And every time we pop CRM out of PPM
we have to reduce it with the removed CRM size. In this PR we will
review and change if we see some issues.