Replicas' TLS key exchange - basic implementation  (#1923)
In this PR we introduce a basic implementation for replicas' TLS key exchange.
A replica TLS key exchange is composed of two phases:

The operator sends a request with up to f replicas that we want to exchange (see below why we don't exchange for more than f replicas at once)
Once the request is executed, each replica in the target_replicas generates a new key-pair and certificate and sends another BFT request that contains the new certificate.
On this second request execution, the replica saves the new certificate and closes the old TLS channel to the exchanged replica.
Note:
We have a basic problem with TLS replicas' TLS key exchange: While the replicas in the lead quorum replace the key on the same stage, the clients need to fetch the data from the replicas in order to exchange the replicas' key.
However, once the replicas changed their TLS keys, the clients cannot connect to the replica and fetch the data.
To overcome the problem we don't allow to rotate more than f replicas at once. This way, both, CRE and ST are still able to work with the remaining quorum. It is the user's responsibility to verify that the clients were able to get the replicas key exchange command before exchanging more keys.
Note, that in case we have lagged/ST replicas, it is the client's responsibility to include these replicas in the exchanged f replicas (otherwise, we won't have a functional quorum).

To complete the feature we need to implement also the following:

ST handler for this command
Client's reconfiguration handler for this command