Introduce Concord Client's StateSnaptshotService
The `StateSnapshotService` can be used to initialize a client
application from a state snapshot visible to Concord Client.

Details and explanations are given as comments inside the protobuf file.

This interface is the output of work mostly done by @meiersi-da,
@andrewjstone and @robem. I have done minor tweaks and naming changes
only.

One of the changes that is worth mentioning is that we allow empty
values when streaming key-values. Rationale is that it is a valid
value and we don't need to concern ourslves here with the special
handling of empty values between the execution engine (EE) and Concord.
Yet, I am open to discuss that also on the EE-Concord interface.