Coverity bug fix
PreProcessBatchRequestMsg and ClientBatchRequestMsg contains a batch of client messages and each message is parsed from the message body for requestLen, signatureLen etc. Theses len parameters are used a offset to point to the message buffer. Coverity tool complains that sanity/range check is not done on these len parameters and they are used in allocation of buffer for these message in PreProcessBatchRequestMsg::getPreProcessRequestMsgs and ClientBatchRequestMsg::getClientPreProcessRequestMsgs.
This PR, adds a function to validate these len parameters