Changes in Concord Client request service
This patch applies a minor change to the Client Request gRPC service
imposed by the new execution engine interface.
The type of request field (in Request message) and response field (in
Response message) are changed from bytes to protobuf.Any.

This change allows a type to be attached to the response field so that
the user of the service can interpret the data easier.