Add more documentation on ClientsManager behavior
This commit adds more documentation to
bftengine/src/bftengine/ClientsManager.hpp with the intention of better
documenting its intended behavior.

This commit is motivated in part by an intention to improve unit testing of the
ClientsManager class; the hope is that this documentation will clarify what
behavior the unit tests should cover.