stabilizing skvbc_batch_preexecution apollo test
In some intermittent failure cases, batch execution is not finished on all replicas when primary is stopped. As a result, new requests are ignored. So, we need to send new request for view change. Adding this check to make sure that, batch execution is finished before we stop primary to test view change