EventGroup: Update tracing
This change propagates the trace context from concord to the caller of concord
client. Thereby, we leverage the existing legacy interface (bytes) for internal
trace context transfer (via thin_replica.proto) but expose it in W3C HTTP
header format within a "text map".
Note, the functionality has not been end-to-end tested given that we still lack
a tracer implementation (such as Jaeger). This will be introduced soon and
should uncover any bugs.