Introduce internal client reconfiguration engine  for state transfer (#1964)
In this PR we introduce an internal client reconfiguration engine for state transfer and read-only replicas.

The state transfer-based CRE has a critical limitation, the replica must complete the state transfer cycle in order to catch up with the state change. However, there are some reconfiguration commands that may prevent a replica to complete state transfer (such as TLS key exchange and scaling).
In addition, with very long state transfer cycles we may want to have an immediate update without waiting for state transfer compilation.
For that, we need to have an async reconfiguration mechanism that runs in parallel to state transfer.

In this PR we introduce the following:

1. For regular replica, once state transfer is starting, an internal CRE is also started.
2. For read-only replica, CRE is always is running.

To demonstrate the ability of this new approach, we implemented a TLS key exchange for state-transferred replicas using this mechanism.
Finally, we added a test for TLS key exchange with a read-only replica.