Docker update to version 0.35
Add Jaeger client. This is a straight port from our closed-source project. The
Jaeger client project will be deprecated very soon and we need to move to
opentelemetry. The upgrade needs to be in sync with closed-source and hence we
use the jaeger-client-cpp instead of opentelemetry-cpp here.
Jaeger depends on thrift protocol which we pull in as an additional dependency.
Note, this should be easier with opentelemetry in the future.
The main motivation for this change is clientservice in order to properly
export trace data.