Clientservice: Add Jaeger
In the clientservice executable, this change initializes Jaeger for tracing. In
the future, we want to move to, probably, opentelemetry but we comply with our
closed-source setup for now.
This change has been tested manually and we could see traces from concord and
the clientservice in a "follow from" relationship in the Jaeger UI. Also, the
clientservice properly returns traces in W3C format in the EventService.