WIP: Introducing CRE handler for scaling during state transfer (#1988)
In this PR we address a fundamental issue with reconfiguration commands that change the network topology, remove metadata and/or restart.

Our initial approach was to rely on state transfer to get an update on such changes, while the leading quorum, commits the relevant update to the chain.
The main problem with this approach is that if the network's topology has changed, we cannot assume anything about the state-transfer's ability to complete the state transfer cycle (and hence get the update).
For example, if we scaled down from 7 to 4, the replica is state transfer is unable to complete the state transfer cycle (it needs at least 5 replicas for state transfer).

To solve this issue, we designed the state transfer CRE for executing reconfiguration commands asynchronously to state transfer compilation (see 89484a6).
However, with scaling the procedure is a bit more complicated due to the following reasons:

We cannot execute the scaling command twice - meaning, that we need some way to mark which scaling command has already been executed.
In order to be able to start after the scaling, we need to remove the metadata, however, this may bring the replica to an inconsistent state once it completes the state transfer (mainly due to changes in reserved pages).
To overcome these issues we do the following:

We save in a dedicated file the new configuration ids. The reason why we don't use the same approach as a regular client (and save an update on the chain) is that the leading quorum will be probably wedged at this time and the state-transferred replica will not be able to update its status on the chain.
On state transfer compilation, the replica detects if there is some new configuration block (at this point, if there was any scaling command it was already executed by the replica's CRE). If there is such an update, the replica removes its metadata and restart itself in order to be consistent with the other replicas.
This PR also contains an Apollo test that demonstrates how this should work.
Notice that read-only replica scaling is not part of this PR.