Add more unit tests for ClientsManager.
This commit adds additional unit tests for the ClientsManager class with
the intention of reaching unit test coverage of as much of
ClientsManager's expected behaviors as possible.

Additionally, this commit also includes some minor fixes to
ClientsManager's implementation to address bugs in it that came up while
writing these unit tests and some minor revisions to the documentation
of ClientManager's intended behavior in its header file to address
certain ambiguities and inaccuracies in it that developing these unit
tests highlighted.