CS/CC: Read the certificates from files at the CS instead of CC
This commit moves the functionality of reading the TLS certs
for TRC-TRS connection to the client service from the concord client.
The helper functions, namely, `decryptPK`, `readCert`,
`getClientIdFromClientCert` and `parseClientIdFromSubject`, have also
been moved to CS configuration code.
Additionally, this commit extends `SubscribeConfig` to hold
TRC ID obtained from the OU field in the client certificate, and
an unordered map with replica ID as key and buffer with the replica's
PEM encoded certificates as value. The former is required to
match the given client_id for correct client authorization at TRS,
and the latter is used to hold pem cert chains for TRS.