Exchange the client tls certs of a server (#2012)
In our current setup, each server has two certificates being used. One when it serves as serve and one when it serves as a client (the specific logic is: A server is a client for the servers with id smaller than its own and server to the rest)
In this PR, we adjust the TLS key rotation logic to this and fix some bugs and stabilize it.