CC/CS: Revert root_cert_chain_map to event_pem_certs in TransportConfig
This commit reverts the map of replicaId -> TRS TLS cert to TransportConfig
because the certs associated with the server endpoints are independent of
SubscribeConfig. A subscription may use unique subscription ID, as well as
unique certs, but the certs associated with the servers should remain
common among all subscriptions.
Consequently, the configuration of the now transport config
`event_pem_certs` is separated into another configureTransport function.
Note, we don't need a map to store the certs, because server.cert is
a composite cert file consisting of all the server certs.