CID correction for preprepare message
The batched CID is set same as seq number.
But when we build Preprepare message in a
separate thread, we assign a temporary seq num
to the PPM (PrePrepareMsg). This seq num is
used to assign the cid to PPM. Later during
consensus, the primary replica assigns the
actual seq num and during that time cid
should be updated.
Another point to note here, if the temporary
cid allocated (based on seq num) is 999
and the actual seq num is 1003, then it will
result into a problem as the space reserved for
cid is 3. To avoid this problem we will put
the CID as 0999 and then fill 1003.
So this will result into addition of an extra
0 for other cases, which should be fine.