Add provision to set and throw exceptions as part of BasicUpdateQueue
This commit updates the definition of BasicUpdateQueue::tryPop() s.t.,
if the exception_ptr is set, it rethrows the exception object pointed
by the exception_ptr. It also adds a setException method, which when
called sets the exception pointer to point to the exception object
passed as an argument to setException.
The objective behind this change is to allow propagation of exceptions
from the TRC to the event_service via teh uodate queue, since the
update_queue is common among all the three components of event_service,
concord_client and the TRC.