Fix race condition between receiveUpdates and updates_queue Pop
Currently BasicUpdateQueue::pop() waits in an endless while
loop until the queue is no longer empty or release_consumers_
has been set.
If receiveUpdates throws an exception while pop() is waiting,
it will not catch the exception. This commit updates
BasicUpdateQueue::pop() s.t., it breaks from the while loop
when the exception is set by receiveUpdatesWrapper, so that
the exception can be rethrown.